+++ pkg/reduct/msg/bucket.py
@@ -3,7 +3,7 @@
 from enum import Enum
 from typing import Optional, List, Dict
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 
 class QuotaType(Enum):
@@ -108,12 +108,6 @@ class QueryEntry(BaseModel):
 
     stop: Optional[int] = None
     """end time in microseconds"""
-
-    include: Dict[str, str] = Field({}, deprecated="Use when instead")
-    """include labels. Added for backward compatibility"""
-
-    exclude: Dict[str, str] = Field({}, deprecated="Use when instead")
-    """exclude labels. Added for backward compatibility"""
 
     each_s: Optional[float] = None
     """return a record every S seconds"""